--- src/AbstractObject/GraphicObject.java
@@ -1,5 +1,0 @@
-package AbstractObject;
-
-public abstract class GraphicObject {
-    public abstract double area();
-}
