--- src/Employee/Manager.java
@@ -1,12 +1,0 @@
-package Employee;
-
-public class Manager extends Employee{
-
-    public Manager(int hours) {
-        super(hours);
-    }
-
-    public double getSalary(){
-        return super.getSalary() * 1.5;
-    }
-}
