--- src/Square_Circle/FigureMetrics.java
@@ -1,6 +1,0 @@
-package Square_Circle;
-
-public interface FigureMetrics {
-    public double calculateArea();
-    public double calculatePerimeter();
-}
