+++ src/abstract_object/GraphicObject.java
@@ -1,0 +1,5 @@
+package abstract_object;
+
+public abstract class GraphicObject {
+    public abstract double area();
+}
