+++ src/employee/Manager.java
@@ -1,0 +1,12 @@
+package employee;
+
+public class Manager extends Employee{
+
+    public Manager(int hours) {
+        super(hours);
+    }
+
+    public double getSalary(){
+        return super.getSalary() * 1.5;
+    }
+}
