+++ src/square_circle/FigureMetrics.java
@@ -1,0 +1,6 @@
+package square_circle;
+
+public interface FigureMetrics {
+    public double calculateArea();
+    public double calculatePerimeter();
+}
